# quizonline-frontend/src/app/api/generated/model/domain-write.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DomainWriteDto { 
    allowed_languages: Array<number>;
    active?: boolean;
    owner?: number;
    managers: Array<number>;
}