# quizonline-server/openapi.yaml
openapi: 3.0.3
info:
  title: QuizOnline API
  version: 1.0.0
paths:
...
  /api/domain/{domain_id}/details/:
    get:
      operationId: domain_details_retrieve
      summary: Récupérer un domaine avec détails
      parameters:
      - in: path
        name: domain_id
        schema:
          type: integer
        description: ID du domaine.
        required: true
      tags:
      - Domain
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainDetail'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Not found
  /api/domain/{domain_id}/member-role/:
    post:
      operationId: domain_member_role_create
      parameters:
      - in: path
        name: domain_id
        schema:
          type: integer
        description: A unique integer value identifying this domain.
        required: true
      tags:
      - domain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainWriteRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/DomainWriteRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DomainWriteRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainWrite'
          description: ''
  /api/domain/available-for-linking/:
    get:
      operationId: domain_available_for_linking_retrieve
      tags:
      - domain
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainWrite'
          description: ''
...
components:
  schemas:
...
    DomainDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        translations:
          allOf:
          - $ref: '#/components/schemas/LocalizedNameDescriptionTranslations'
          readOnly: true
        allowed_languages:
          type: array
          items:
            $ref: '#/components/schemas/LanguageRead'
          readOnly: true
        active:
          type: boolean
          readOnly: true
        owner:
          allOf:
          - $ref: '#/components/schemas/UserSummary'
          readOnly: true
        managers:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/SubjectRead'
          readOnly: true
      required:
      - active
      - allowed_languages
      - created_at
      - id
      - managers
      - owner
      - subjects
      - translations
      - updated_at
    DomainNameSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
      - id
      - name
...
    DomainWrite:
      type: object
      properties:
        allowed_languages:
          type: array
          items:
            type: integer
        active:
          type: boolean
        owner:
          type: integer
        managers:
          type: array
          items:
            type: integer
      required:
      - allowed_languages
      - managers
    DomainWriteRequest:
      type: object
      properties:
        translations:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/LocalizedNameDescriptionTranslationRequest'
          writeOnly: true
        allowed_languages:
          type: array
          items:
            type: integer
        active:
          type: boolean
        owner:
          type: integer
        managers:
          type: array
          items:
            type: integer
      required:
      - allowed_languages
      - managers
      - translations
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
...
    LanguageRead:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        code:
          type: string
          readOnly: true
          description: 'Code ISO (ex: fr, nl, en, fr-BE)'
        name:
          type: string
          readOnly: true
          description: 'Nom lisible (ex: Français, Nederlands, English)'
        active:
          type: boolean
          readOnly: true
          description: Langue disponible pour la saisie et l'affichage
      required:
      - active
      - code
      - id
      - name
...
    LocalizedNameDescriptionTranslation:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
      - description
      - name
    LocalizedNameDescriptionTranslationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
      required:
      - description
      - name
    LocalizedNameDescriptionTranslations:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/LocalizedNameDescriptionTranslation'
...
    LocalizedSubjectTranslation:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        domain:
          $ref: '#/components/schemas/DomainNameSummary'
      required:
      - description
      - domain
      - name
    LocalizedSubjectTranslations:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/LocalizedSubjectTranslation'
...
    SubjectRead:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        domain:
          type: integer
          readOnly: true
        active:
          type: boolean
          readOnly: true
        translations:
          allOf:
          - $ref: '#/components/schemas/LocalizedSubjectTranslations'
          readOnly: true
      required:
      - active
      - domain
      - id
      - translations
...
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
      required:
      - id
      - username
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT